% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_searching.R
\name{search_gnr}
\alias{search_gnr}
\title{Retrieve and parse full taxonomic information for a list of taxa.}
\usage{
search_gnr(species_list,
  excluded_sources = c("Open Tree of Life Reference Taxonomy",
  "The Paleobiology Database", "Union 4", "Wikispecies"))
}
\arguments{
\item{species_list}{(Char) A vector of taxa names.}

\item{excluded_sources}{(Char) A vector of sources to exclude (see
\code{taxize::gnr_datasources()}. By default, excludes Open Tree of Life Reference
Taxonomy, The Paleobiology Database, Union 4, and Wikispecies because rank output
from these sources is poorly structured and hard to parse.}
}
\value{
A dataframe of taxonomic information, with every rank in a separate column.
}
\description{
\code{search_gnr()} is a wrapper for \code{taxize::gnr_resolve()} that parses its
output and returns a neat dataframe of taxonomic information, with every rank in a
separate column. The ICN ranks (kingdom, phylum, class, order, family, genus) are
ordered so that they can be accessed with \code{kingdom:genus}. All other ranks
(super-, sub-, infra- versions of ICN ranks, plus tribe) are ordered to the right.
Finally, the raw \code{path} and \code{ranks} strings are retained so that you can
double-check them if you want.
}
\examples{
my_species <- c("Coptodactyla meridionalis", "Torymus chrysochlorus", "Anaspis rufa")
search_gnr(my_species)
}
