% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxizehelper.R
\docType{package}
\name{taxizehelper}
\alias{taxizehelper}
\alias{taxizehelper-package}
\title{Convenience functions for working with the \code{taxize} package.}
\description{
\code{taxizehelper} is a personal package by Desi Quintans. It contains functions that I've
developed to help me use \code{taxize} functions efficiently in my Rmarkdown documents.
}
\section{Author}{

Desi Quintans (@eco_desi)
}

\section{URL}{

\url{https://github.com/DesiQuintans/taxizehelper/}
}

