% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wsu_hwk_insects}
\alias{wsu_hwk_insects}
\title{The insects of Western Sydney University, Hawkesbury Campus}
\format{A character vector with 28 entries: \describe{ \item{[1]}{"Apis mellifera",
  "Heteronympha merope", "Yoyetta celis", ...} }}
\source{
\url{https://www.ala.org.au}
}
\usage{
wsu_hwk_insects
}
\description{
A list of insect species recorded within 1 km of Western Sydney University, Hawkesbury
Campus (NSW, Australia).
}
\keyword{datasets}
